pyproject.toml
[tool.poetry]
name = "k1core-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.8"
django = "^5.1.6"
uvicorn = "^0.34.0"
celery = "^5.4.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^6.0.0"

[tool.isort]
profile = "black"
skip = ["migrations",]

[tool.flake8]
exclude = ["migrations", ]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
